requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
djangorestframework>=3.11.0,<3.12.0
flake8>=3.6.0,<3.7.8
django-cors-headers
Pillow>=5.3.0,<5.4.8
